refactor: db로 네이밍 변경
data -> db
