+++ app/build.gradle.kts
@@ -59,6 +59,8 @@ dependencies {
     implementation(libs.google.material)
     implementation(libs.androidx.activity)
     implementation("androidx.fragment:fragment-ktx:1.8.6")
+    implementation("androidx.room:room-runtime:2.6.1")
+    kapt("androidx.room:room-compiler:2.6.1")
     testImplementation(libs.assertj.core)
     testImplementation(libs.junit.jupiter)
     testImplementation(libs.kotest.runner.junit5)
