+++ app/src/main/java/woowacourse/movie/db/DummyMovie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.db
 
 import woowacourse.movie.R
 import woowacourse.movie.domain.model.Movie
